refactor(STADTPULS-385): rename records by id request
